don't allow DELETE methon in other user public folder

+++ src/index.js
@@ -140,6 +140,10 @@ export default {
 			}
 
 			if (request.method === "DELETE") {
+				if (objectName.indexOf(writeRoot) !== 0) {
+					return makeResponse(undefined, { status: 400 })
+				}
+				
 				await env.storage.delete(url.pathname.slice(1))
 				return makeResponse()
 			}
